pipeline/requirements.txt
-e .
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline_actions
aws-cdk.aws-codebuild
aws-cdk.aws-codedeploy
aws-cdk.aws-secretsmanager
aws-cdk.aws-iam
aws-cdk.core
python-dotenv
